Move decompores protocol into dxe_core
